Fixing errors
Fixed some "seeing through wall"problems by adjusting resolution parameters. Changed main code to be easier to adjust matrix resolution. Changed problem where when the guard would be spawned on a wall, it wouldnt be reassigned to a new location, although it painted a new location as having a guard. Fixed problem that made guards not show up in the final diagram.
Lastly, the coordiantes used as input for position of the guard were inverted so in my part of the code the y coordinate was being treated as the x coordinate and vice versa.